requirements.txt
Flask==3.1.1
stripe==12.2.0
requests==2.32.4
openai==1.30.5
langchain==0.1.17
langchain-openai==0.1.7
chromadb==0.4.22
tiktoken==0.7.0
python-dotenv==1.0.1
gunicorn==21.2.0

